Move parts of config.go to new module, adjust tests, remove some error checks
