+++ exporters/metric/cortex/go.mod
@@ -4,8 +4,7 @@ go 1.14
 
 require (
 	github.com/google/go-cmp v0.5.1
-	github.com/spf13/afero v1.3.3
-	github.com/spf13/viper v1.7.1
 	go.opentelemetry.io/otel v0.10.0
 	go.opentelemetry.io/otel/sdk v0.10.0
+	google.golang.org/genproto v0.0.0-20191108220845-16a3f7862a1a // indirect
 )
