+++ exporters/metric/cortex/utils/go.mod
@@ -1,0 +1,13 @@
+module opentelemetry.io/contrib/exporters/metric/cortex/utils
+
+go 1.14
+
+replace opentelemetry.io/contrib/exporters/metric/cortex => ../
+
+require (
+	github.com/google/go-cmp v0.5.1
+	github.com/spf13/afero v1.3.3
+	github.com/spf13/viper v1.7.1
+	github.com/stretchr/testify v1.6.1
+	opentelemetry.io/contrib/exporters/metric/cortex v0.0.0-00010101000000-000000000000
+)
